Simplify bot startup
