--- config.yaml
@@ -1,49 +1,0 @@
-currency_rates:
-  USD: 100.0
-  EUR: 110.0
-  RUB: 1.0
-clearance_fee_ranges:
-  - [200000, 1067]
-  - [450000, 2134]
-  - [1200000, 4269]
-  - [3000000, 11746]
-  - [5000000, 16524]
-  - [7000000, 20000]
-  - [999999999999, 30000]
-tariffs:
-  base_clearance_fee: 3100
-  base_util_fee: 20000
-  ctp_util_coeff_base: 1.2
-  etc_util_coeff_base: 1.0
-  excise_rates:
-    gasoline: 58
-    diesel: 58
-    electric: 0
-    hybrid: 58
-  recycling_factors:
-    default:
-      gasoline: 1.0
-      diesel: 1.0
-      electric: 0.3
-      hybrid: 1.0
-    adjustments:
-      "5-7":
-        gasoline: 0.26
-        diesel: 0.26
-        electric: 0.26
-        hybrid: 0.26
-  age_groups:
-    overrides:
-      "5-7":
-        gasoline:
-          rate_per_cc: 4.8
-          min_duty: 0
-        diesel:
-          rate_per_cc: 5.0
-          min_duty: 0
-        electric:
-          rate_per_cc: 0
-          min_duty: 1000
-        hybrid:
-          rate_per_cc: 2.0
-          min_duty: 2500
